requirements.txt
django_rest_framework
django
celery
django-extensions
python-dateutil
pykafka